Refactor: Move Host-Header Site Naming Into Its Own Module
Ten module-level names in api_resource.py existed to turn a Host header
into a display name: a TLD set, two validation patterns, a system
dictionary load, a word splitter and two entry points. None of it is
about routing, SQL, or serving a request.

api/utils/site_name.py, with the 21 tests that cover it moved alongside
into test_site_name.py. The tests that stay in test_api_resource.py are
the two asserting the name is injected into the rendered page, which is
an api_resource concern.

One test patched api.api_resource._WORDS, which no longer exists there.
monkeypatch.setattr raised rather than silently creating it, so the stale
target failed loudly.

api_resource.py: 2506 -> 2423 lines.
