+++ api/api_resource.py
@@ -14,7 +14,6 @@ import pathlib
 import re
 import threading
 import time
-import urllib.parse
 import uuid
 
 # Imported at runtime, not under TYPE_CHECKING, because route handlers annotate parameters with it and
@@ -24,7 +23,7 @@ import uuid
 # unnecessary once handlers carry a route decorator.
 from collections.abc import Sequence  # noqa: TC003
 from datetime import timedelta
-from functools import lru_cache, wraps
+from functools import wraps
 from typing import TYPE_CHECKING, Any
 from typing import cast as typecast
 
@@ -55,6 +54,7 @@ from api.utils.generation_cache import GenerationCache
 from api.utils.http_utils import make_user_agent
 from api.utils.param_binding import ParamCoercionError
 from api.utils.routing import build_route_table, build_routes_listing, route
+from api.utils.site_name import hostname_to_site_name
 from api.utils.timer import Timer
 from card_engine import ENGINE_COLUMNS as _ENGINE_COLUMNS_FROM_MODULE
 from card_engine import QueryEngine as _QueryEngine
@@ -84,8 +84,6 @@ def _rss_mb() -> str:
         pass
     return "unknown"
 
-
-FALLBACK_SITE_NAME = "MTG Search"
 
 # Placeholder written into index.html/card.html wherever the site name belongs, so the substitution
 # below can't accidentally match unrelated copy that happens to contain "MTG Search".
@@ -148,88 +146,6 @@ def _minify_html(html: str) -> str:
     if not _MINIFY_HTML_ENABLED:
         return html
     return minify_html.minify(html, minify_js=True, minify_css=True, keep_comments=True)
-
-
-# TLDs in this set are stripped from the hostname; others are concatenated into the word.
-# e.g. sylvan-librarian.com -> "Sylvan Librarian"; tolarian-acade.my -> "Tolarian Academy"
-_STRIP_TLDS = frozenset(["app", "biz", "co", "com", "dev", "edu", "gov", "info", "io", "me", "net", "org", "us"])
-# Allowlist: only valid hostname characters (letters, digits, hyphens, dots) after urlparse extracts the host.
-_SAFE_HOSTNAME_RE = re.compile(r"^[a-z0-9.\-]+$")
-_IP_RE = re.compile(r"^\d+\.\d+\.\d+\.\d+$")
-
-
-_MIN_WORD_LEN = 3
-
-
-# TODO: supplement with words from https://api.scryfall.com/catalog/word-bank so MtG-specific
-# terms (e.g. "sylvan", "planeswalker") are recognized even on systems without a full dictionary.
-def _load_word_set() -> frozenset[str]:
-    for path in ("/usr/share/dict/american-english", "/usr/share/dict/words"):
-        try:
-            with pathlib.Path(path).open() as f:
-                return frozenset(w.lower() for w in f.read().splitlines() if w.isalpha() and len(w) >= _MIN_WORD_LEN)
-        except OSError:
-            continue
-    return frozenset()
-
-
-_WORDS: frozenset[str] = _load_word_set()
-
-
-def _split_words(s: str, words: frozenset[str]) -> list[str] | None:
-    """Split s into dictionary words by searching split points from the middle outward.
-
-    Returns a list of words on success, or None if the string cannot be fully partitioned.
-    """
-    # Only attempt dictionary splitting for purely alphabetic strings.
-    if not s.isalpha():
-        return None
-
-    @lru_cache(maxsize=4096)
-    def _split(sub: str) -> tuple[str, ...] | None:
-        if sub in words:
-            return (sub,)
-        n = len(sub)
-        if n < _MIN_WORD_LEN:
-            return None
-        mid = n // 2
-        for k in sorted(range(1, n), key=lambda k: abs(k - mid)):
-            left, right = sub[:k], sub[k:]
-            if left in words:
-                rest = _split(right)
-                if rest is not None:
-                    return (left, *rest)
-            if right in words:
-                rest = _split(left)
-                if rest is not None:
-                    return (*rest, right)
-        return None
-
-    res = _split(s)
-    return list(res) if res is not None else None
-
-
-def hostname_to_site_name(raw_host: str) -> str:
-    """Derive a display name from a Host header value, falling back to FALLBACK_SITE_NAME."""
-    # urlparse requires a scheme; .hostname strips the port and lowercases.
-    hostname = urllib.parse.urlparse(f"http://{raw_host}").hostname or ""
-    return _hostname_to_site_name(hostname[:64])
-
-
-@lru_cache(maxsize=256)
-def _hostname_to_site_name(hostname: str) -> str:
-    if not hostname or hostname == "localhost" or _IP_RE.match(hostname) or not _SAFE_HOSTNAME_RE.match(hostname):
-        return FALLBACK_SITE_NAME
-    parts = hostname.split(".")[-2:]
-    tld = parts[-1].lower()
-    name = parts[0] if tld in _STRIP_TLDS else ".".join(parts)
-    name = name.replace(".", "").replace("-", " ").strip()
-    if " " not in name and _WORDS:
-        split = _split_words(name, _WORDS)
-        if split is not None:
-            name = " ".join(split)
-    name = name.title()
-    return name if any(c.isalnum() for c in name) else FALLBACK_SITE_NAME
 
 
 # Query parameters that must not be forwarded to action handlers.
